@@ -11,15 +11,25 @@
   },
   "license": "MIT",
   "files": [
-    "index.js",
-    "lib"
+    "dist"
   ],
-  "main": "index.js",
+  "type": "module",
+  "main": "dist/index.cjs",
+  "module": "dist/index.js",
+  "exports": {
+    ".": {
+      "require": "./dist/index.cjs",
+      "import": "./dist/index.js"
+    }
+  },
   "engines": {
-    "node": ">=8.6"
+    "node": ">=12"
   },
   "scripts": {
-    "lint": "eslint --cache --cache-location node_modules/.cache/.eslintcache --report-unused-disable-directives --ignore-path .gitignore .",
+    "build": "tsup lib/index.js --format=esm,cjs --clean",
+    "lint": "run-p lint:*",
+    "lint:lib": "eslint ./lib/**/*.js --fix",
+    "lint:test": "eslint ./test/**/*.js --fix",
     "mocha": "mocha --reporter dot",
     "test": "npm run lint && npm run mocha",
     "test:ci": "npm run test:cover",
@@ -29,9 +39,13 @@
     "eslint": "^6.8.0",
     "fill-range": "^7.0.1",
     "gulp-format-md": "^2.0.0",
-    "mocha": "^6.2.2",
-    "nyc": "^15.0.0",
-    "time-require": "github:jonschlinkert/time-require"
+    "mocha": "^10.0.0",
+    "npm-run-all": "^4.1.5",
+    "nyc": "^15.1.0",
+    "time-require": "github:jonschlinkert/time-require",
+    "ts-node": "^10.9.1",
+    "tsup": "^6.1.3",
+    "typescript": "^4.7.4"
   },
   "keywords": [
     "glob",
