@@ -1,19 +1,15 @@
-'use strict';
 
-const constants = require('./constants');
-const utils = require('./utils');
+import * as utils from './utils.js';
 
-/**
- * Constants
- */
-
-const {
+import {
   MAX_LENGTH,
   POSIX_REGEX_SOURCE,
   REGEX_NON_SPECIAL_CHARS,
   REGEX_SPECIAL_CHARS_BACKREF,
-  REPLACEMENTS
-} = constants;
+  REPLACEMENTS,
+  globChars,
+  extglobChars
+} from './constants.js';
 
 /**
  * Helpers
@@ -52,7 +48,7 @@ const syntaxError = (type, char) => {
  * @return {Object}
  */
 
-const parse = (input, options) => {
+export const parse = (input, options) => {
   if (typeof input !== 'string') {
     throw new TypeError('Expected a string');
   }
@@ -74,8 +70,8 @@ const parse = (input, options) => {
   const win32 = utils.isWindows(options);
 
   // create constants based on platform, for windows or posix
-  const PLATFORM_CHARS = constants.globChars(win32);
-  const EXTGLOB_CHARS = constants.extglobChars(PLATFORM_CHARS);
+  const PLATFORM_CHARS = globChars(win32);
+  const EXTGLOB_CHARS = extglobChars(PLATFORM_CHARS);
 
   const {
     DOT_LITERAL,
@@ -1023,7 +1019,7 @@ parse.fastpaths = (input, options) => {
     NO_DOTS_SLASH,
     STAR,
     START_ANCHOR
-  } = constants.globChars(win32);
+  } = globChars(win32);
 
   const nodot = opts.dot ? NO_DOTS : NO_DOT;
   const slashDot = opts.dot ? NO_DOTS_SLASH : NO_DOT;
@@ -1087,5 +1083,3 @@ parse.fastpaths = (input, options) => {
 
   return source;
 };
-
-module.exports = parse;
