@@ -1,10 +1,9 @@
-'use strict';
+import path from 'path';
+import { scan } from './scan.js';
+import { parse } from './parse.js';
+import * as utils from './utils.js';
+import * as constants from './constants.js';
 
-const path = require('path');
-const scan = require('./scan');
-const parse = require('./parse');
-const utils = require('./utils');
-const constants = require('./constants');
 const isObject = val => val && typeof val === 'object' && !Array.isArray(val);
 
 /**
@@ -28,7 +27,6 @@ const isObject = val => val && typeof val === 'object' && !Array.isArray(val);
  * @return {Function=} Returns a matcher function.
  * @api public
  */
-
 const picomatch = (glob, options, returnState = false) => {
   if (Array.isArray(glob)) {
     const fns = glob.map(input => picomatch(input, options, returnState));
@@ -113,7 +111,6 @@ const picomatch = (glob, options, returnState = false) => {
  * @return {Object} Returns an object with matching info.
  * @api public
  */
-
 picomatch.test = (input, regex, options, { glob, posix } = {}) => {
   if (typeof input !== 'string') {
     throw new TypeError('Expected input to be a string');
@@ -157,7 +154,6 @@ picomatch.test = (input, regex, options, { glob, posix } = {}) => {
  * @return {Boolean}
  * @api public
  */
-
 picomatch.matchBase = (input, glob, options, posix = utils.isWindows(options)) => {
   const regex = glob instanceof RegExp ? glob : picomatch.makeRe(glob, options);
   return regex.test(path.basename(input));
@@ -179,7 +175,6 @@ picomatch.matchBase = (input, glob, options, posix = utils.isWindows(options)) =
  * @return {Boolean} Returns true if any patterns match `str`
  * @api public
  */
-
 picomatch.isMatch = (str, patterns, options) => picomatch(patterns, options)(str);
 
 /**
@@ -195,7 +190,6 @@ picomatch.isMatch = (str, patterns, options) => picomatch(patterns, options)(str
  * @return {Object} Returns an object with useful properties and output to be used as a regex source string.
  * @api public
  */
-
 picomatch.parse = (pattern, options) => {
   if (Array.isArray(pattern)) return pattern.map(p => picomatch.parse(p, options));
   return parse(pattern, { ...options, fastpaths: false });
@@ -227,7 +221,6 @@ picomatch.parse = (pattern, options) => {
  * @return {Object} Returns an object with
  * @api public
  */
-
 picomatch.scan = (input, options) => scan(input, options);
 
 /**
@@ -241,7 +234,6 @@ picomatch.scan = (input, options) => scan(input, options);
  * @return {RegExp}
  * @api public
  */
-
 picomatch.compileRe = (state, options, returnOutput = false, returnState = false) => {
   if (returnOutput === true) {
     return state.output;
@@ -338,5 +330,4 @@ picomatch.constants = constants;
 /**
  * Expose "picomatch"
  */
-
-module.exports = picomatch;
+export default picomatch;
