@@ -1,7 +1,7 @@
-'use strict';
 
-const utils = require('./utils');
-const {
+import * as utils from './utils.js';
+
+import {
   CHAR_ASTERISK,             /* * */
   CHAR_AT,                   /* @ */
   CHAR_BACKWARD_SLASH,       /* \ */
@@ -17,7 +17,7 @@ const {
   CHAR_RIGHT_CURLY_BRACE,    /* } */
   CHAR_RIGHT_PARENTHESES,    /* ) */
   CHAR_RIGHT_SQUARE_BRACKET  /* ] */
-} = require('./constants');
+} from './constants.js';
 
 const isPathSeparator = code => {
   return code === CHAR_FORWARD_SLASH || code === CHAR_BACKWARD_SLASH;
@@ -46,7 +46,7 @@ const depth = token => {
  * @api public
  */
 
-const scan = (input, options) => {
+export const scan = (input, options) => {
   const opts = options || {};
 
   const length = input.length - 1;
@@ -388,4 +388,3 @@ const scan = (input, options) => {
   return state;
 };
 
-module.exports = scan;
