@@ -1,27 +1,26 @@
-'use strict';
 
-const path = require('path');
-const win32 = process.platform === 'win32';
-const {
+import path from 'path';
+import {
   REGEX_BACKSLASH,
   REGEX_REMOVE_BACKSLASH,
   REGEX_SPECIAL_CHARS,
   REGEX_SPECIAL_CHARS_GLOBAL
-} = require('./constants');
+} from './constants.js';
+const win32 = process.platform === 'win32';
 
-exports.isObject = val => val !== null && typeof val === 'object' && !Array.isArray(val);
-exports.hasRegexChars = str => REGEX_SPECIAL_CHARS.test(str);
-exports.isRegexChar = str => str.length === 1 && exports.hasRegexChars(str);
-exports.escapeRegex = str => str.replace(REGEX_SPECIAL_CHARS_GLOBAL, '\\$1');
-exports.toPosixSlashes = str => str.replace(REGEX_BACKSLASH, '/');
+export const isObject = val => val !== null && typeof val === 'object' && !Array.isArray(val);
+export const hasRegexChars = str => REGEX_SPECIAL_CHARS.test(str);
+export const isRegexChar = str => str.length === 1 && hasRegexChars(str);
+export const escapeRegex = str => str.replace(REGEX_SPECIAL_CHARS_GLOBAL, '\\$1');
+export const toPosixSlashes = str => str.replace(REGEX_BACKSLASH, '/');
 
-exports.removeBackslashes = str => {
+export const removeBackslashes = str => {
   return str.replace(REGEX_REMOVE_BACKSLASH, match => {
     return match === '\\' ? '' : match;
   });
 };
 
-exports.supportsLookbehinds = () => {
+export const supportsLookbehinds = () => {
   const segs = process.version.slice(1).split('.').map(Number);
   if (segs.length === 3 && segs[0] >= 9 || (segs[0] === 8 && segs[1] >= 10)) {
     return true;
@@ -29,21 +28,22 @@ exports.supportsLookbehinds = () => {
   return false;
 };
 
-exports.isWindows = options => {
+export const isWindows = options => {
   if (options && typeof options.windows === 'boolean') {
     return options.windows;
   }
   return win32 === true || path.sep === '\\';
 };
 
-exports.escapeLast = (input, char, lastIdx) => {
+export const escapeLast = (input, char, lastIdx) => {
   const idx = input.lastIndexOf(char, lastIdx);
   if (idx === -1) return input;
-  if (input[idx - 1] === '\\') return exports.escapeLast(input, char, idx - 1);
-  return `${input.slice(0, idx)}\\${input.slice(idx)}`;
+  return input[idx - 1] === '\\'
+    ? escapeLast(input, char, idx - 1)
+    : input.slice(0, idx) + '\\' + input.slice(idx);
 };
 
-exports.removePrefix = (input, state = {}) => {
+export const removePrefix = (input, state = {}) => {
   let output = input;
   if (output.startsWith('./')) {
     output = output.slice(2);
@@ -52,7 +52,7 @@ exports.removePrefix = (input, state = {}) => {
   return output;
 };
 
-exports.wrapOutput = (input, state = {}, options = {}) => {
+export const wrapOutput = (input, state = {}, options = {}) => {
   const prepend = options.contains ? '' : '^';
   const append = options.contains ? '' : '$';
 
