@@ -1,9 +1,9 @@
-'use strict';
 
-const assert = require('assert');
-const fill = require('fill-range');
-const match = require('./support/match');
-const { isMatch } = require('..');
+import assert from 'assert';
+import fill from 'fill-range';
+import match from './support/match.js';
+import picomatch from '../lib/index.js';
+const { isMatch } = picomatch;
 
 describe('braces', () => {
   it('should not match with brace patterns when disabled', () => {
