@@ -1,8 +1,8 @@
-'use strict';
 
-const assert = require('assert');
-const match = require('./support/match');
-const { isMatch } = require('..');
+import assert from 'assert';
+import match from './support/match.js';
+import picomatch from '../lib/index.js';
+const { isMatch } = picomatch;
 
 describe('dotfiles', () => {
   describe('normal', () => {
@@ -224,7 +224,7 @@ describe('dotfiles', () => {
       assert(isMatch('abc/../abc', '*/../*'));
     });
 
-    it('should not match double dots when not defined in pattern', async() => {
+    it('should not match double dots when not defined in pattern', async () => {
       assert(!isMatch('../abc', '**/*'));
       assert(!isMatch('../abc', '**/**/**'));
       assert(!isMatch('../abc', '**/**/abc'));
@@ -291,7 +291,7 @@ describe('dotfiles', () => {
       assert(!isMatch('abc/abc/..', 'abc/*/**/*', { strictSlashes: true }));
     });
 
-    it('should not match single exclusive dots when not defined in pattern', async() => {
+    it('should not match single exclusive dots when not defined in pattern', async () => {
       assert(!isMatch('.', '**'));
       assert(!isMatch('abc/./abc', '**'));
       assert(!isMatch('abc/abc/.', '**'));
