@@ -1,8 +1,8 @@
-'use strict';
 
-const assert = require('assert');
-const match = require('./support/match');
-const { isMatch, makeRe } = require('..');
+import assert from 'assert';
+import match from './support/match.js';
+import picomatch from '../lib/index.js';
+const { isMatch, makeRe } = picomatch;
 
 /**
  * Ported from Bash 4.3 and 4.4 unit tests
