@@ -1,8 +1,7 @@
-'use strict';
 
-const assert = require('assert');
-const match = require('./support/match');
-const picomatch = require('..');
+import assert from 'assert';
+import match from './support/match.js';
+import picomatch from '../lib/index.js';
 const { isMatch } = picomatch;
 
 const equal = (actual, expected, msg) => {
