@@ -1,9 +1,8 @@
-'use strict';
 
-const assert = require('assert');
-const pm = require('..');
-const { makeRe, parse } = pm;
-
+import assert from 'assert';
+import pm from '../lib/index.js';
+import { parse } from '../lib/parse.js';
+const { makeRe } = pm;
 const opts = { strictSlashes: true, posix: true, regex: true };
 const isMatch = (...args) => pm.isMatch(...args, opts);
 const convert = (...args) => {
