@@ -1,6 +1,5 @@
-'use strict';
 
-const pm = require('..');
+const pm = require('../lib');
 
 const isMatch = pm.matcher('*', { ignore: 'f*' });
 console.log(isMatch('foo')); //=> false
